requirements-test.txt
apache-airflow
wixflow
pytest
aiohttp
requests
boto3
PyJWT
pyspark
Jinja2
sqlparse
setuptools